# swagger/swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "ShopEase API",
      "version": "1.0.0",
      "description": "API Documentation for ShopEase"
    },
    "servers": [
      {
        "url": "http://localhost:3000",
        "description": "Local server"
      }
    ],
    "paths": {
      "/products": {
        "get": {
          "summary": "Get all products",
          "responses": {
            "200": {
              "description": "A list of products",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Product"
                    }
                  }
                }
              }
            }
          }
        },
        "post": {
          "summary": "Add a new product",
          "requestBody": {
            "description": "Product object that needs to be added to the store",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            },
            "required": true
          },
          "responses": {
            "201": {
              "description": "Product created",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Product"
                  }
                }
              }
            }
          }
        }
      },
      "/products/{id}": {
        "get": {
          "summary": "Get a specific product by ID",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Product details",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Product"
                  }
                }
              }
            },
            "404": {
              "description": "Product not found"
            }
          }
        },
        "put": {
          "summary": "Update an existing product",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer"
              }
            }
          ],
          "requestBody": {
            "description": "Product object that needs to be updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "Product updated",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Product"
                  }
                }
              }
            },
            "404": {
              "description": "Product not found"
            }
          }
        },
        "delete": {
          "summary": "Delete a product",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer"
              }
            }
          ],
          "responses": {
            "204": {
              "description": "Product deleted"
            },
            "404": {
              "description": "Product not found"
            }
          }
        }
      }
    },
    "components": {
      "schemas": {
        "Product": {
          "type": "object",
          "properties": {
            "id": {
              "type": "integer",
              "example": 1
            },
            "name": {
              "type": "string",
              "example": "Product name"
            },
            "description": {
              "type": "string",
              "example": "Product description"
            },
            "price": {
              "type": "number",
              "format": "float",
              "example": 19.99
            },
            "inventory": {
              "type": "integer",
              "example": 100
            }
          },
          "required": ["name", "description", "price"]
        }
      }
    }
  }
  